Organize and disable preprocessed image endpoints
